adding review(read)

+++ src/main/java/com/TubesDiKaosan/ecommerce/models/Riviews.java
@@ -22,13 +22,13 @@ public class Riviews {
     @ManyToOne
     @JoinColumn(name = "user_id")
     private Users user;
-    private Integer rate;
+    private Float rate;
     private String comment;
     
     public Riviews() {
     }
 
-    public Riviews(Integer review_id, Product product, Users user, Integer rate, String comment) {
+    public Riviews(Integer review_id, Product product, Users user, Float rate, String comment) {
         this.review_id = review_id;
         this.product = product;
         this.user = user;
@@ -60,11 +60,11 @@ public class Riviews {
         this.user = user;
     }
 
-    public Integer getRate() {
+    public Float getRate() {
         return rate;
     }
 
-    public void setRate(Integer rate) {
+    public void setRate(Float rate) {
         this.rate = rate;
     }
